# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SafePay API
  version: "1.0.0"
  description: |
    **Trusted payments, everywhere.**

    SafePay is an escrow, settlement and trust-scoring layer for Nigeria. Any
    application — a marketplace, a freelance platform, a storefront, a landlord's
    portal — can plug in so that a transaction between two people who do not fully
    trust each other becomes safe by default.

    ## Authentication

    Two credentials reach the same endpoints:

    | Credential | Header | Used by |
    |---|---|---|
    | Dashboard session (JWT) | `Authorization: Bearer eyJ…` | The SafePay web app |
    | API key | `Authorization: Bearer sk_test_…` | Partner applications |

    Keys come in pairs. `sk_test_` keys operate on simulated settlement — build the
    entire flow, including disputes and webhooks, without moving a naira.
    SafePay stores only a SHA-256 hash of every key, so a lost key is rotated,
    never recovered.

    ## Money

    Every amount crosses the wire as an **integer in kobo** (`amountKobo`). Endpoints
    that create an escrow also accept `amount` in naira as a convenience, and convert
    it. Floating point never touches a balance.

    ## Errors

    Failures return a consistent envelope:

    ```json
    { "error": { "code": "conflict", "message": "An escrow that is \"released\" cannot move to \"released\"." } }
    ```

    ## Rate limits

    Responses carry `RateLimit-Limit`, `RateLimit-Remaining` and `RateLimit-Reset`.
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/escrows:
    post:
      tags: [Escrows]
      summary: Create an escrow
      description: |
        Creates an escrow in `created` state. No money moves until it is funded.

        For `service_milestone`, the milestone amounts must sum exactly to the
        escrow total — SafePay rejects the request otherwise rather than silently
        reconciling the difference.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [goods, service_milestone, rental, recurring, in_person]
                amount:
                  type: number
                  description: Amount in naira. Ignored when `amountKobo` is supplied.
                  example: 185000
                amountKobo:
                  type: integer
                  description: Canonical amount in kobo.
                  example: 18500000
                title: { type: string, example: iPhone 13 Pro, 256GB }
                description: { type: string }
                sellerId: { type: string }
                sellerEmail:
                  type: string
                  format: email
                  description: Used when the counterparty is not yet a SafePay user; they are invited.
                role:
                  type: string
                  enum: [buyer, seller]
                  default: buyer
                  description: Whether the caller is paying or being paid.
                milestones:
                  type: array
                  items:
                    type: object
                    properties:
                      title: { type: string }
                      amount: { type: number }
                      amountKobo: { type: integer }
            examples:
              goods:
                summary: A phone sale
                value:
                  type: goods
                  amount: 185000
                  title: iPhone 13 Pro, 256GB
                  sellerEmail: seller@example.com
              milestones:
                summary: Freelance work in three stages
                value:
                  type: service_milestone
                  amount: 450000
                  title: Brand website build
                  milestones:
                    - { title: Design mockups, amount: 150000 }
                    - { title: Frontend build, amount: 200000 }
                    - { title: CMS and handover, amount: 100000 }
      responses:
        "201":
          description: Escrow created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  escrow: { $ref: "#/components/schemas/Escrow" }
                  flags:
                    type: array
                    description: Fraud-detection labels raised on creation, if any.
                    items: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
    get:
      tags: [Escrows]
      summary: List escrows
      parameters:
        - { name: status, in: query, schema: { type: string }, description: "Comma-separated statuses." }
        - { name: type, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: string, enum: [buyer, seller] } }
      responses:
        "200":
          description: Escrows the caller is party to, with a summary.
          content:
            application/json:
              schema:
                type: object
                properties:
                  escrows:
                    type: array
                    items: { $ref: "#/components/schemas/Escrow" }
                  summary: { $ref: "#/components/schemas/EscrowSummary" }
  /v1/escrows/{id}:
    get:
      tags: [Escrows]
      summary: Retrieve an escrow with its ledger
      parameters: [{ $ref: "#/components/parameters/EscrowId" }]
      responses:
        "200":
          description: The escrow and its append-only money trail.
          content:
            application/json:
              schema:
                type: object
                properties:
                  escrow: { $ref: "#/components/schemas/Escrow" }
                  ledger:
                    type: array
                    items: { $ref: "#/components/schemas/LedgerEntry" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/escrows/{id}/fund:
    post:
      tags: [Escrows]
      summary: Fund an escrow
      description: Buyer only. Moves `created` → `funded` and starts the auto-release clock.
      parameters: [{ $ref: "#/components/parameters/EscrowId" }]
      responses:
        "200": { $ref: "#/components/responses/EscrowResponse" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /v1/escrows/{id}/release:
    post:
      tags: [Escrows]
      summary: Release the funds
      description: |
        Buyer only, and **irreversible**. Pays the seller, collects the platform fee,
        contributes to the Buyer Protection Reserve, and recalculates both parties'
        SafeScores.
      parameters: [{ $ref: "#/components/parameters/EscrowId" }]
      responses:
        "200": { $ref: "#/components/responses/EscrowResponse" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/escrows/{id}/cancel:
    post:
      tags: [Escrows]
      summary: Cancel an unfunded escrow
      parameters: [{ $ref: "#/components/parameters/EscrowId" }]
      responses:
        "200": { $ref: "#/components/responses/EscrowResponse" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/escrows/{id}/milestones/{milestoneId}/approve:
    post:
      tags: [Escrows]
      summary: Approve and release one milestone
      description: Buyer only. Releases that portion; approving the last one settles the escrow.
      parameters:
        - { $ref: "#/components/parameters/EscrowId" }
        - { name: milestoneId, in: path, required: true, schema: { type: string }, example: ms_2 }
      responses:
        "200": { $ref: "#/components/responses/EscrowResponse" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/escrows/claim:
    post:
      tags: [Escrows]
      summary: Claim an in-person escrow by code
      description: The buyer scans the seller's QR (or types the short code) to join the escrow.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, example: KRT4-9WQP }
      responses:
        "200": { $ref: "#/components/responses/EscrowResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/disputes:
    post:
      tags: [Disputes]
      summary: Raise a dispute
      description: |
        Freezes the escrow and classifies the complaint before any human reads it —
        category, severity, confidence and a suggested next step for the reviewer.
        Uses Gemini when configured, and a deterministic rule-based classifier
        otherwise, so the flow never depends on a third party being up.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [escrowId, reason]
              properties:
                escrowId: { type: string }
                reason: { type: string, minLength: 12 }
                evidenceUrls:
                  type: array
                  items: { type: string, format: uri }
      responses:
        "201":
          description: Dispute opened and triaged.
          content:
            application/json:
              schema:
                type: object
                properties:
                  dispute: { $ref: "#/components/schemas/Dispute" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      tags: [Disputes]
      summary: List disputes
      description: Regular accounts see their own; admins see everything.
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [open, under_review, resolved] } }
      responses:
        "200":
          description: Disputes visible to the caller.
          content:
            application/json:
              schema:
                type: object
                properties:
                  disputes:
                    type: array
                    items: { $ref: "#/components/schemas/Dispute" }
  /v1/disputes/{id}/resolve:
    post:
      tags: [Disputes]
      summary: Resolve a dispute
      description: Admin only. Settles the escrow and updates both SafeScores.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome]
              properties:
                outcome:
                  type: string
                  enum: [release_to_seller, refund_buyer]
                note: { type: string }
                coverFromReserve:
                  type: boolean
                  description: Pay the refund out of the Buyer Protection Reserve.
      responses:
        "200":
          description: Resolved.
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A dashboard session JWT, or an API key beginning `sk_test_` / `sk_live_`.
  parameters:
    EscrowId:
      name: id
      in: path
      required: true
      schema: { type: string }
      example: esc_Xk4VrVTBEROWlXi7
  responses:
    EscrowResponse:
      description: The updated escrow.
      content:
        application/json:
          schema:
            type: object
            properties:
              escrow: { $ref: "#/components/schemas/Escrow" }
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Forbidden:
      description: Authenticated, but not permitted to do this.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: The resource is not in a state that allows this transition.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: conflict }
            message: { type: string }
            details: { type: object }
...
    Party:
      type: object
      nullable: true
      properties:
        id: { type: string, nullable: true }
        name: { type: string }
        safeScore: { type: integer }
        scoreTier: { type: string }
        invited: { type: boolean, description: Present when the counterparty has no SafePay account yet. }
    Milestone:
      type: object
      properties:
        id: { type: string, example: ms_1 }
        title: { type: string }
        amountKobo: { type: integer }
        status: { type: string, enum: [pending, approved] }
        approvedAt: { type: string, format: date-time, nullable: true }
    Escrow:
      type: object
      properties:
        id: { type: string, example: esc_Xk4VrVTBEROWlXi7 }
        type: { type: string, enum: [goods, service_milestone, rental, recurring, in_person] }
        title: { type: string }
        description: { type: string }
        status:
          type: string
          enum: [created, funded, in_progress, released, disputed, refunded, cancelled, expired]
        amountKobo: { type: integer, example: 18500000 }
        feeKobo: { type: integer, example: 277500 }
        netToSellerKobo: { type: integer, example: 18222500 }
        currency: { type: string, example: NGN }
        buyer: { $ref: "#/components/schemas/Party" }
        seller: { $ref: "#/components/schemas/Party" }
        milestones:
          type: array
          nullable: true
          items: { $ref: "#/components/schemas/Milestone" }
        claimCode: { type: string, nullable: true, example: KRT4-9WQP }
        flagged: { type: boolean }
        autoReleaseAt: { type: string, format: date-time, nullable: true }
        fundedAt: { type: string, format: date-time, nullable: true }
        releasedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        timeline:
          type: array
          items:
            type: object
            properties:
              event: { type: string }
              at: { type: string, format: date-time }
              note: { type: string, nullable: true }
    EscrowSummary:
      type: object
      properties:
        total: { type: integer }
        inEscrowKobo: { type: integer }
        releasedKobo: { type: integer }
        earnedKobo: { type: integer }
        openDisputes: { type: integer }
        awaitingAction: { type: integer }
    LedgerEntry:
      type: object
      description: Append-only. A correction is a new, opposite entry — never an edit.
      properties:
        id: { type: string }
        escrowId: { type: string }
        type: { type: string, enum: [fund, release, refund, fee, reserve, reserve_payout] }
        amountKobo: { type: integer }
        note: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    Triage:
      type: object
      description: Produced automatically the moment a dispute is filed.
      properties:
        category:
          type: string
          enum: [non_delivery, not_as_described, service_incomplete, partial_delivery, damaged, likely_fraud, buyer_remorse, other]
        label: { type: string, example: Likely fraud }
        severity: { type: string, enum: [low, medium, high, critical] }
        confidence: { type: number, format: float, example: 0.77 }
        summary: { type: string }
        guidance: { type: string, example: Freeze the counterparty account and escalate to compliance before refunding. }
        source: { type: string, example: "rules", description: "`rules` or `gemini:<model>`." }
    Dispute:
      type: object
      properties:
        id: { type: string }
        escrowId: { type: string }
        raisedById: { type: string }
        raisedByRole: { type: string, enum: [buyer, seller] }
        againstId: { type: string }
        reason: { type: string }
        evidenceUrls:
          type: array
          items: { type: string, format: uri }
        ai: { $ref: "#/components/schemas/Triage" }
        status: { type: string, enum: [open, under_review, resolved] }
        resolution:
          type: object
          nullable: true
          properties:
            outcome: { type: string, enum: [release_to_seller, refund_buyer] }
            note: { type: string }
        createdAt: { type: string, format: date-time }
...